+++ xiao-round-rig-controller-v-2-1/src/storage/ConfigStore.h
@@ -1,18 +1,7 @@
 #pragma once
-
 #include "model/RigTypes.h"
-
 namespace rig {
-
-class ConfigStore {
- public:
-  StorageBootResult loadAtBoot(RigConfig &config);
-
- private:
-  bool loadConfigFile(RigConfig &config);
-  bool writeTemplate();
-  void appendBootEvent(const char *event, const char *detail = "");
-  void assignValue(RigConfig &config, const String &key, const String &value);
-};
-
-}  // namespace rig
+enum class ConfigField:uint8_t{WifiSsid,WifiPassword,ApiBase,NodeId,BearerToken};
+class ConfigStore {public:StorageBootResult loadAtBoot(RigConfig&);bool saveOverride(ConfigField,const char*,RigConfig&);bool clearOverrides();
+ private:bool loadConfigFile(RigConfig&);bool loadOverrides(RigConfig&);bool writeTemplate();void appendBootEvent(const char*,const char*="");void assignValue(RigConfig&,const String&,const String&);};
+}